+++ dotnet/Models/AffirmCaptureRequest.cs
@@ -6,6 +6,9 @@
 
     public class AffirmCaptureRequest
     {
+        public int amount { get; set; }
         public string order_id { get; set; }
+        public string shipping_carrier { get; set; }
+        public string shipping_confirmation { get; set; }
     }
 }